# docs/swagger.yaml
definitions:
...
  models.ResTokenImg:
    properties:
      large:
        type: string
      ownImg:
        type: string
      small:
        type: string
      thumb:
        type: string
      tokenAddress:
        type: string
      tokenChainId:
        type: integer
    type: object
  models.ResTokenList:
    properties:
      chainId:
        type: integer
      contractAddress:
        type: string
      daoName:
        items:
          type: string
        type: array
      tokenAddress:
        type: string
      totalSupply:
        type: string
    type: object
  models.ResTokenListPage:
    properties:
      list:
        items:
          $ref: '#/definitions/models.ResTokenList'
        type: array
      total:
        type: integer
    type: object
...
  models.ResVotesList:
    properties:
      amount:
        type: string
      optionIndex:
        type: integer
      proposalId:
        type: integer
      voter:
        type: string
    type: object
  models.ResVotesListPage:
    properties:
      list:
        items:
          $ref: '#/definitions/models.ResVotesList'
        type: array
      total:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /stpdao/v2/token/img:
    get:
      description: query Token img
      parameters:
      - description: tokenChainId
        in: query
        name: tokenChainId
        required: true
        type: integer
      - description: tokenAddress
        in: query
        name: tokenAddress
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResTokenImg'
      summary: query Token img
      tags:
      - Token
  /stpdao/v2/token/list:
    get:
      description: query Token list
      parameters:
      - description: chainId
        in: query
        name: chainId
        type: string
      - description: creator
        in: query
        name: creator
        type: string
      - description: offset,page
        in: query
        name: offset
        required: true
        type: integer
      - description: count,page
        in: query
        name: count
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResTokenListPage'
      summary: query Token list
      tags:
      - Token
  /stpdao/v2/votes/list:
    get:
      description: query votes list
      parameters:
      - description: chainId
        in: query
        name: chainId
        required: true
        type: integer
      - description: proposalId
        in: query
        name: proposalId
        required: true
        type: string
      - description: Dao address
        in: query
        name: daoAddress
        required: true
        type: string
      - description: offset,page
        in: query
        name: offset
        required: true
        type: integer
      - description: count,page
        in: query
        name: count
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResVotesListPage'
      summary: query votes list
      tags:
      - votes
swagger: "2.0"